+++ aws-distro-opentelemetry-node-autoinstrumentation/src/instrumentation/common/instrumentation-utils.ts
@@ -69,6 +69,8 @@ export interface AttributeMapping {
   from: string;
   to?: string;
   transform?: (value: any, attrs: Record<string, any>) => any;
+  // Apply the transformation even when the destination attribute already exists.
+  overrideDestinationIfExists?: boolean;
 }
 
 export function tryParseJson(value: string): unknown {